+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/OidcSiopVerifier.kt
@@ -287,7 +287,10 @@ class OidcSiopVerifier(
     /**
      * Creates an OIDC Authentication Request, encoded as query parameters to the [walletUrl].
      */
-    @Deprecated("Replace with new class", ReplaceWith("createAuthnRequest(requestOptions, OpenId4VpVerifier.CreationOptions.Query(walletUrl)).getOrThrow().url"))
+    @Deprecated(
+        "Replace with new class",
+        ReplaceWith("createAuthnRequest(requestOptions, OpenId4VpVerifier.CreationOptions.Query(walletUrl)).getOrThrow().url")
+    )
     suspend fun createAuthnRequestUrl(
         walletUrl: String,
         requestOptions: RequestOptions,
@@ -302,7 +305,10 @@ class OidcSiopVerifier(
      * Creates an OIDC Authentication Request, encoded as query parameters to the [walletUrl],
      * containing a JWS Authorization Request (JAR, RFC9101) in `request`, containing the request parameters itself.
      */
-    @Deprecated("Replace with new class", ReplaceWith("createAuthnRequest(requestOptions, OpenId4VpVerifier.CreationOptions.SignedRequestByValue(walletUrl)).url"))
+    @Deprecated(
+        "Replace with new class",
+        ReplaceWith("createAuthnRequest(requestOptions, OpenId4VpVerifier.CreationOptions.SignedRequestByValue(walletUrl)).url")
+    )
     suspend fun createAuthnRequestUrlWithRequestObject(
         walletUrl: String,
         requestOptions: RequestOptions,
@@ -325,7 +331,10 @@ class OidcSiopVerifier(
      * @param requestUrl the URL where the request itself can be loaded by the client
      * @return The URL to display to the Wallet, and the JWS that shall be made accessible under [requestUrl]
      */
-    @Deprecated("Replace with new class", ReplaceWith("createAuthnRequest(requestOptions, OpenId4VpVerifier.CreationOptions.SignedRequestByReference(walletUrl, requestUrl))"))
+    @Deprecated(
+        "Replace with new class",
+        ReplaceWith("createAuthnRequest(requestOptions, OpenId4VpVerifier.CreationOptions.SignedRequestByReference(walletUrl, requestUrl))")
+    )
     suspend fun createAuthnRequestUrlWithRequestObjectByReference(
         walletUrl: String,
         requestUrl: String,
@@ -354,7 +363,10 @@ class OidcSiopVerifier(
      * // on an GET to requestUrl, return `jar.serialize()`
      * ```
      */
-    @Deprecated("Replace with new class", ReplaceWith("createAuthnRequest(requestOptions, OpenId4VpVerifier.CreationOptions.SignedRequestByReference(walletUrl, requestUrl))"))
+    @Deprecated(
+        "Replace with new class",
+        ReplaceWith("createAuthnRequest(requestOptions, OpenId4VpVerifier.CreationOptions.SignedRequestByReference(walletUrl, requestUrl))")
+    )
     suspend fun createAuthnRequestAsSignedRequestObject(
         requestOptions: RequestOptions,
     ): KmmResult<JwsSigned<AuthenticationRequestParameters>> = catching {
@@ -788,12 +800,10 @@ class OidcSiopVerifier(
         val sessionTranscript = SessionTranscript(
             deviceEngagementBytes = null,
             eReaderKeyBytes = null,
-            handover = ByteStringWrapper(
-                OID4VPHandover(
-                    clientIdHash = clientIdToHash.serialize().sha256(),
-                    responseUriHash = responseUriToHash.serialize().sha256(),
-                    nonce = challenge
-                )
+            handover = OID4VPHandover(
+                clientIdHash = clientIdToHash.serialize().sha256(),
+                responseUriHash = responseUriToHash.serialize().sha256(),
+                nonce = challenge
             ),
         )
         val deviceAuthentication = DeviceAuthentication(