+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/PresentationFactory.kt
@@ -8,8 +8,9 @@ import at.asitplus.dif.PresentationDefinition
 import at.asitplus.openid.*
 import at.asitplus.signum.indispensable.CryptoPublicKey
 import at.asitplus.signum.indispensable.cosef.CoseSigned
-import at.asitplus.signum.indispensable.cosef.io.Base16Strict
 import at.asitplus.signum.indispensable.cosef.io.ByteStringWrapper
+import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
+import at.asitplus.signum.indispensable.io.Base64Strict
 import at.asitplus.signum.indispensable.josef.JsonWebKey
 import at.asitplus.signum.indispensable.josef.JwsSigned
 import at.asitplus.signum.indispensable.josef.toJsonWebKey
@@ -24,6 +25,7 @@ import io.matthewnelson.encoding.base16.Base16
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
 import kotlinx.datetime.Clock
 import kotlinx.serialization.builtins.ByteArraySerializer
+import kotlinx.serialization.encodeToByteArray
 import kotlin.coroutines.cancellation.CancellationException
 import kotlin.random.Random
 import kotlin.time.Duration.Companion.seconds
@@ -93,11 +95,16 @@ internal class PresentationFactory(
         responseUrl: String?,
         nonce: String,
         docType: String,
-    ): Pair<CoseSigned<ByteArray>, String?> = if (responseWillBeEncrypted && clientId != null && responseUrl != null) {
+    ): Pair<CoseSigned<ByteArray>, String?> = if (clientId != null && responseUrl != null) {
         val deviceNameSpaceBytes = ByteStringWrapper(DeviceNameSpaces(mapOf()))
-        val mdocGeneratedNonce = Random.Default.nextBytes(16).encodeToString(Base16Strict)
-        val clientIdToHash =
-            ClientIdToHash(clientId = clientId, mdocGeneratedNonce = mdocGeneratedNonce)
+        // if it's not encrypted, we have no way of transporting the mdocGeneratedNonce,
+        // so we'll use the empty string
+        val mdocGeneratedNonce = if (responseWillBeEncrypted)
+            Random.Default.nextBytes(16).encodeToString(Base64Strict) else ""
+        val clientIdToHash = ClientIdToHash(
+            clientId = clientId,
+            mdocGeneratedNonce = mdocGeneratedNonce
+        )
         val responseUriToHash = ResponseUriToHash(
             responseUri = responseUrl,
             mdocGeneratedNonce = mdocGeneratedNonce
@@ -105,12 +112,10 @@ internal class PresentationFactory(
         val sessionTranscript = SessionTranscript(
             deviceEngagementBytes = null,
             eReaderKeyBytes = null,
-            handover = ByteStringWrapper(
-                OID4VPHandover(
-                    clientIdHash = clientIdToHash.serialize().sha256(),
-                    responseUriHash = responseUriToHash.serialize().sha256(),
-                    nonce = nonce
-                )
+            handover = OID4VPHandover(
+                clientIdHash = clientIdToHash.serialize().sha256(),
+                responseUriHash = responseUriToHash.serialize().sha256(),
+                nonce = nonce
             ),
         )
         val deviceAuthentication = DeviceAuthentication(
@@ -119,9 +124,14 @@ internal class PresentationFactory(
             docType = docType,
             namespaces = deviceNameSpaceBytes
         )
-        Napier.d("Device authentication is ${deviceAuthentication.serialize().encodeToString(Base16())}")
+        // TODO Change type to CoseSigned<DeviceAuthentication>
+        val deviceAuthenticationBytes = coseCompliantSerializer
+            .encodeToByteArray(ByteStringWrapper(deviceAuthentication))
+            .wrapInCborTag(24)
+            .also { Napier.d("Device authentication signature input is ${it.encodeToString(Base16())}") }
+
         coseService.createSignedCoseWithDetachedPayload(
-            payload = deviceAuthentication.serialize(),
+            payload = deviceAuthenticationBytes,
             serializer = ByteArraySerializer(),
             addKeyId = false
         ).getOrElse {